pyproject.toml
[project]
name = "ragnar"
version = "0.1.1"
requires-python = ">=3.10"
description = "RAG utilities and app backend for CanaR (Insee)"
readme = "README.md"

dependencies = [
  "qdrant-client>=1.9.0",
  "markdown-it-py>=3.0.0",
  "PyYAML>=6.0.1",
  "numpy>=1.26.0",
  "requests>=2.31.0",
  "tiktoken>=0.7.0",
  "python-dotenv>=1.0",
  "tqdm>=4.66",
]

[project.scripts]
ragnar-ingest = "ragnar.ingest_utilitr:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["ragnar*"]